# packages/server/src/modules/user/user.entity.ts
import {
  Entity,
  Column,
  JoinColumn,
  PrimaryColumn,
  CreateDateColumn,
  UpdateDateColumn,
  ManyToOne,
} from 'typeorm';
import { ApiProperty } from '@nestjs/swagger';
import {
  GenderEnum, RoleEnum, BotLanguageEnum, CountriesEnum,
} from '@common_bot/shared';

@Entity('users')
export class UserEntity {
  @ApiProperty({
    example: '266006070',
  })
  @PrimaryColumn({ type: 'varchar' })
  id: string;

  @ApiProperty({
    example: 'Ivan',
    nullable: true,
  })
  @Column({
    type: 'varchar',
    nullable: true,
  })
  firstname?: string;

  @ApiProperty({
    example: 'Ivanov',
    nullable: true,
  })
  @Column({
    type: 'varchar',
    nullable: true,
  })
  lastname?: string;

  @ApiProperty({
    example: 'sergei_zheludkov',
    nullable: true,
  })
  @Column({
    type: 'varchar',
    unique: true,
    nullable: true,
  })
  username?: string;

  @ApiProperty({
    enum: BotLanguageEnum,
  })
  @Column({
    type: 'enum',
    enum: BotLanguageEnum,
    default: BotLanguageEnum.ENGLISH,
  })
  lang: BotLanguageEnum;

  @ApiProperty({
    enum: CountriesEnum,
  })
  @Column({
    type: 'enum',
    enum: CountriesEnum,
  })
  country: CountriesEnum;

  @ApiProperty({
    enum: RoleEnum,
  })
  @Column({
    type: 'enum',
    enum: RoleEnum,
    default: RoleEnum.USER,
  })
  role: RoleEnum;

  @ApiProperty({
    enum: GenderEnum,
  })
  @Column({
    type: 'enum',
    enum: GenderEnum,
  })
  gender: GenderEnum;

  @ApiProperty({
    // eslint-disable-next-line no-use-before-define
    type: UserEntity,
    nullable: true,
  })
  @ManyToOne(() => UserEntity, {
    nullable: true,
  })
  @JoinColumn({
    name: 'who_invited_id',
  })
  who_invited?: string;

  @ApiProperty({
    example: 5,
  })
  @Column({
    type: 'int',
    default: 0,
  })
  referral_counter: number;

  @ApiProperty({
    example: 5,
  })
  @Column({
    type: 'int',
    default: 0,
  })
  complete_tasks_count: number;

  @ApiProperty({
    example: 0.000_000_1,
  })
  @Column({
    type: 'decimal',
    default: 0.0,
  })
  mining_rate: number;

  @ApiProperty({
    example: 0.5,
  })
  @Column({
    type: 'decimal',
    default: 0.0,
  })
  withdrawn_tons: number;

  @ApiProperty({
    example: 0.5,
  })
  @Column({
    type: 'decimal',
    default: 0.0,
  })
  balance: number;

  @ApiProperty({
    example: '2022-10-21T19:48:59.726Z',
    nullable: true,
  })
  @Column({
    type: 'timestamptz',
    nullable: true,
  })
  mining_rate_started: Date;

  @ApiProperty({
    example: '2022-10-21T19:48:59.726Z',
  })
  @CreateDateColumn()
  created: Date;

  @ApiProperty({
    example: '2022-10-21T19:48:59.726Z',
  })
  @UpdateDateColumn()
  updated: Date;
}